v1.1.2: Developer Experience
## Highlights
- Ruff Phase 2: 9,706 → 3,188 violations (−67%)
- 2,280 unsafe-fixes applied, 8 new rule ignores
- justfile: 20+ targets (modern task runner)
- .devcontainer: GitHub Codespaces ready
- Pre-commit: bandit → ty + uv-lock hooks

## New files
- justfile
- .devcontainer/devcontainer.json

## Modified
- pyproject.toml: version 1.1.2, 8 new ruff ignores
- .pre-commit-config.yaml: ty hook, uv-lock hook, removed bandit
- CHANGELOG.md: v1.1.2 entry
- src/codomyrmex/__init__.py: version 1.1.2
- ~65 source files: ruff unsafe-fixes applied
- All root docs: version sync v1.1.2

+++ src/codomyrmex/bio_simulation/ant_colony/ant.py
@@ -112,7 +112,7 @@ class Ant:
         for idx, pos in enumerate(reversed(self.pheromone_trail)):
             decay = strength * (1.0 - idx / max(trail_len, 1))
             if decay > 0:
-                grid_pos = (int(round(pos[0])), int(round(pos[1])))
+                grid_pos = (round(pos[0]), round(pos[1]))
                 deposits.append((grid_pos, decay))
 
         return deposits
